tagcounter/requirements.txt
amqp==2.5.0
aniso8601==6.0.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
drf-extensions==0.5.0
entrypoints==0.3
flake8==3.7.8
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
kombu==4.6.3
mccabe==0.6.1
mock==3.0.5
promise==2.2.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
redis==3.3.4
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
